requirements.txt
absl-py==1.0.0
aiohttp==3.8.1
aiosignal==1.2.0
aksharamukha==2.0.1
altair==4.2.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.0
blinker==1.4
cachetools==5.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
conllu==4.4.2
cryptography==37.0.2
cycler==0.11.0
datasets==2.1.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.4
docker-pycreds==0.4.0
elastic-transport==8.1.2
elasticsearch==8.2.0
entrypoints==0.4
executing==0.8.3
faiss-cpu==1.7.2
fastjsonschema==2.15.3
filelock==3.6.0
fire==0.4.0
flatbuffers==1.12
fonttools==4.33.3
frozenlist==1.3.0
fsspec==2022.3.0
future==0.18.2
gast==0.4.0
git-python==1.0.3
gitdb==4.0.9
gitdb2==4.0.2
GitPython==3.0.8
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
googleapis-common-protos==1.56.1
grpcio==1.46.3
h11==0.13.0
h5py==3.7.0
huggingface-hub==0.5.1
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
indic-transliteration==2.3.31
iniconfig==1.1.1
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jaconv==0.3
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.2
langcodes==3.3.0
language-data==1.1
libclang==14.0.1
lxml==4.9.0
marisa-trie==0.7.7
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==6.0.2
multiprocess==0.70.12.2
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nlp==0.2.0
nltk==3.7
notebook==6.4.11
numpy==1.22.3
oauthlib==3.2.0
opt-einsum==3.3.0
outcome==1.1.0
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pluggy==1.0.0
portalocker==2.4.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydeck==0.7.1
pyDeprecate==0.3.2
pyenchant==3.2.2
Pygments==2.12.0
pykakasi==2.2.1
Pympler==1.0.1
pyOpenSSL==22.0.0
pyparsing==3.0.8
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
pytorch-lightning==1.0.4
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==23.0.0
regex==2022.4.24
requests==2.27.1
requests-oauthlib==1.3.1
responses==0.18.0
roman==3.3
rouge-score==0.0.4
rsa==4.8
sacrebleu==2.0.0
sacremoses==0.0.49
scikit-learn==1.1.1
scipy==1.8.1
selenium==4.2.0
semver==2.13.0
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.5.12
seqeval==1.2.2
setproctitle==1.2.3
shortuuid==1.0.9
simpletransformers==0.63.7
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.2.0
streamlit==1.9.0
tabulate==0.8.9
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-datasets==4.5.2
tensorflow-estimator==2.9.0
tensorflow-gpu==2.9.1
tensorflow-io-gcs-filesystem==0.26.0
tensorflow-metadata==1.8.0
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
torch==1.8.0
torchmetrics==0.8.2
torchsummary==1.5.1
tornado==6.1
tqdm==4.64.0
traitlets==5.2.1.post0
-e git+https://github.com/fairwell-coding/nlp.git@041e19466f266af158a07eb8d2bb6078f8e5b354#egg=transformers&subdirectory=distillation_test/transformers-main
trio==0.20.0
trio-websocket==0.9.2
typer==0.4.1
typing_extensions==4.2.0
tzdata==2022.1
tzlocal==4.2
unicodedata2==14.0.0
urllib3==1.26.9
validators==0.19.0
wandb==0.12.16
watchdog==2.1.8
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.2
widgetsnbextension==3.6.0
wrapt==1.14.1
wsproto==1.1.0
xxhash==3.0.0
yarl==1.7.2
zipp==3.8.0
